1.0.8.8.1
Refined swapping between the two characters. The spirit character should
exist +5 unities on the 'z' axis from now on in order to function
properly. Same goes for any platforms he is to collide with. The human
player remains unchanged.

Refined platform turning on and off. Box collider remains intact when
swapping now. Platforms are turn on or off on 'Start()' depending on if
they are a spirit or human platform.