1.0.8.9
Attempted to refine jumping, but it isn't working quite like I would
have hoped. Jumping seems to be random, even though it's using constant
values and multiplied by Time.deltaTime. Unsure of a fix at the moment.

The ability to jump when colliding with the side of an object was added
as well, but velocity needs to be added to the player's horizontal
movement upon jumping. This needs work to implement.

Refined character animation when walking. It is now based on his current
velocity. Seems smooth enough.

Added a base custom editor. Has the ability to group objects in a
collapsible field. Decently customizable, but each value much be input
manually. No known workaround is known right now.